+++ CHANGELOG.md
@@ -33,7 +33,7 @@ for further information about branching and tagging conventions.
 - For sales that were ended show highlighted text like a "Closed [days] ago"
 - Reword create opportunity label "KYC Required" => "Verification required"
 - Asset loading via the loop where was `filter[limit]=100`
-- Correct calculating of the base/quote price for Opportunities
+- Correct base/quote price calculating for Opportunities
 
 ### Changed
 - Now using new account endpoint for loading the fees